+++ src/lib.rs
@@ -33,9 +33,7 @@
 
 extern crate alloc;
 
-use ::core::ops::Deref;
 use codec::{Decode, Encode, MaxEncodedLen};
-use core::convert::TryFrom;
 use fixed::types::I64F64;
 
 pub use crate::error::GeohashError;
@@ -70,51 +68,6 @@ static BASE32_CODES: &[char] = &[
     MaxEncodedLen,
 )]
 pub struct GeoHash<const LEN: usize>(pub [u8; LEN]);
-
-impl<const LEN: usize> Deref for GeoHash<LEN> {
-    type Target = [u8; LEN];
-
-    fn deref(&self) -> &Self::Target {
-        &self.0
-    }
-}
-
-impl<const LEN: usize> TryFrom<&str> for GeoHash<LEN> {
-    type Error = GeohashError;
-
-    fn try_from(value: &str) -> Result<Self, Self::Error> {
-        TryFrom::<&[u8]>::try_from(value.as_bytes())
-    }
-}
-
-impl<const LEN: usize> TryFrom<[u8; LEN]> for GeoHash<LEN> {
-    type Error = GeohashError;
-
-    fn try_from(value: [u8; LEN]) -> Result<Self, Self::Error> {
-        TryFrom::<&[u8]>::try_from(&value[..])
-    }
-}
-
-impl<const LEN: usize> TryFrom<&[u8]> for GeoHash<LEN> {
-    type Error = GeohashError;
-
-    fn try_from(value: &[u8]) -> Result<Self, Self::Error> {
-        // `try_from` is only successful if the input is a valid base 32 encoded geo hash.
-
-        if value.len() != LEN {
-            return Err(GeohashError::InvalidLen);
-        }
-
-        for c in value.iter() {
-            let _ = hash_value_of_char(*c as char)?;
-        }
-
-        let mut arr = [0u8; LEN];
-        arr.clone_from_slice(value);
-
-        Ok(GeoHash(arr))
-    }
-}
 
 impl<const LEN: usize> GeoHash<LEN> {
     /// Internal function to encode a coordinate to a geohash with length `LEN`.
@@ -373,5 +326,6 @@ fn hash_value_of_char(c: char) -> Result<usize, GeohashError> {
     Err(GeohashError::InvalidHashCharacter(c))
 }
 
+mod convert;
 mod error;
 mod neighbors;
